my-jekyll-theme/tools/requirements.txt
PyYAML>=5.1
Markdown>=3.0
requests>=2.0
